# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  friends.ActivityItem:
    properties:
      album_art_url:
        type: string
      album_name:
        type: string
      artist_name:
        type: string
      avatar_id:
        type: string
      id:
        type: string
      is_playing:
        type: boolean
      platform:
        type: string
      started_at:
        type: string
      track_name:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
...
  friends.Friend:
    properties:
      avatar_id:
        type: string
      compatibility:
        type: integer
      current_track:
        $ref: '#/definitions/friends.ActivityItem'
      id:
        type: string
      is_online:
        type: boolean
      username:
        type: string
    type: object
...
  notifications.NotificationActor:
    properties:
      avatar_id:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  notifications.NotificationItem:
    properties:
      actor:
        $ref: '#/definitions/notifications.NotificationActor'
      body:
        type: string
      created_at:
        type: string
      entity_id:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      read_at:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  notifications.NotificationListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/notifications.NotificationItem'
        type: array
      unreadCount:
        type: integer
    type: object
...
  spotify.ActivityItem:
    properties:
      album_art_url:
        type: string
      album_name:
        type: string
      artist_name:
        type: string
      avatar_id:
        type: string
      id:
        type: string
      is_playing:
        type: boolean
      platform:
        type: string
      preview_url:
        type: string
      spotify_url:
        type: string
      started_at:
        type: string
      track_id:
        type: string
      track_name:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  spotify.ConnectRequest:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
    required:
    - access_token
    - expires_in
    - refresh_token
    type: object
host: localhost:5000
info:
  contact: {}
  description: API documentation for the Chen, a music compatibility and social platform.
  title: Chen API
  version: "1.0"
paths:
...
  /friends:
    get:
      description: Fetch the current user's friend list with current activity and
        compatibility scores
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/friends.Friend'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get Friends List
      tags:
      - friends
...
  /notifications:
    get:
      description: Fetch all notifications for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notifications.NotificationListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List Notifications
      tags:
      - notifications
...
  /spotify/connect:
    post:
      consumes:
      - application/json
      description: Connect or refresh Spotify connection with access tokens
      parameters:
      - description: Spotify connection tokens
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/spotify.ConnectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Connect Spotify Account
      tags:
      - spotify
  /spotify/now-playing:
    get:
      description: Fetch the currently playing track from user's Spotify
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/spotify.ActivityItem'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get Currently Playing Track
      tags:
      - spotify
  /spotify/recent:
    get:
      description: Fetch user's recently played tracks from Spotify
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/spotify.ActivityItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get Recent Tracks
      tags:
      - spotify
  /spotify/top-artists:
    get:
      description: Fetch user's top artists from Spotify
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties:
                type: string
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get Top Artists
      tags:
      - spotify
  /spotify/top-tracks:
    get:
      description: Fetch user's top tracks from Spotify
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/spotify.ActivityItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get Top Tracks
      tags:
      - spotify
swagger: "2.0"